requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pytube3==9.6.4